Components / Facet (change): changes to facet-filters and facet.service for customization of displayed facets
Access user facets directly from the service and make the customization of facets optional.
Legacy "enableAddFacet" has been removed, we are now using the "selected" parameter of the Action to check if a facet is added or removed.

facet.service --> add two function to set the "allFacets" and "defaultFacets" properties in case it was not set in the providers