🔧 ES-22939 (config) Add conditional startConfig initialization based on isDevMode()
The code changes in app.module.ts files introduce a conditional initialization of the startConfig object based on the isDevMode() function. This allows for different configurations depending on whether the application is running in development mode or not. By doing this, the startConfig object can be customized for different environments, such as setting the app name and production flag. This change improves the flexibility and configurability of the application.