ES-26667 #done Handle array values in Utils.toJson() function
Handle array values in `toJson()` to prevent unintended processing of non-string inputs. If an array is passed, it will now be returned directly.
---
Update `proxy.conf.local.json` to be an array of configurations, which is the expected format for Angular CLI proxy setups. This ensures the local proxy works correctly.